Added a Image Dictionary for Image Viewing
Basically, it harvests all your image tags from the HTML string you give
the ImageDictionary class. It then gets all img tags, and assembles a
jeditorpane with the image directory, a colon, then the actual image
picture.

I included a main method for quick testing. It works - Feel free to put
whatever HTML in the string with img tags and try it out!

Formatting isn't pretty. so yea, fix it if you want.