Enhanced save, save as, and close tab
if you try and close a tab that has not saved changes it will prompt you
to either save changes, ignore changes, or abort closing the tab. Also
if you try to execute a save on a tab that has never been saved before,
the program will automatically rexecute a save as, show a JFileChooser
and allow the user to save the file just as if they chose save as in the
first place. Total time 45 minutes.